Add common ModelWrapper
